feat: add enemies attacks

+++ src/entities/players/AbstractPlayer.py
@@ -1,7 +1,8 @@
 import pygame
 from abc import ABC, abstractmethod
-from config.Constants import Constants
-from src.entities.Projectile import ProjectileGenerator
+from config.Constants import Constants, Colors
+from entities.projectiles.ProjectileGenerator import ProjectileGenerator
+from entities.projectiles.BaseProjectile import BaseProjectile
 from src.entities.Abilitiy import MissileBarrage
 from src.entities.Abilitiy import LaserBeam
 from src.entities.Abilitiy import CriticalShot
@@ -45,7 +46,8 @@ class AbstractPlayer(pygame.sprite.Sprite, ABC):
                                                         self.get_projectile_speed(),
                                                         self.get_projectile_frequency(),
                                                         projectile_image,
-                                                        self.get_projectile_damage())
+                                                        self.get_projectile_damage(),
+                                                        is_player_projectile=True)
 
         ability_image = pygame.Surface(
             (Constants.ABILITY_WIDTH, Constants.ABILITY_HEIGHT)
@@ -238,11 +240,16 @@ class AbstractPlayer(pygame.sprite.Sprite, ABC):
         Computes projectile collision and damage taken.
         :param enemies_projectiles: Enemies projectiles on screen.
         """
-
         for projectile in enemies_projectiles:
             if pygame.sprite.collide_rect(self, projectile):
-                self._health_points -= projectile.damage
-                projectile.kill()
+                # Se for uma bomba, só causa dano se ainda não explodiu
+                if hasattr(projectile, '_exploded'):
+                    if not projectile._exploded:
+                        self._health_points -= projectile.damage
+                        projectile._explode(None, self)
+                else:
+                    self._health_points -= projectile.damage
+                    projectile.kill()
 
     def _compute_cooldown_ability(self, dt):
         """
